Fix libpq because it was reading from the backend as a short.

+++ src/interfaces/libpq/libpq-fe.h
@@ -78,7 +78,7 @@ extern		"C"
 		char	   *name;		/* type name */
 		Oid			adtid;		/* type id */
 		short		adtsize;	/* type size */
-		short		adtmod;		/* type-specific modifier info */
+		int			adtmod;		/* type-specific modifier info */
 	} PGresAttDesc;
 
 /* use char* for Attribute values,
